refactor header view
