+++ js/views/footer-view.js
