+++ js/views/header-view.js
@@ -1,26 +1,18 @@
 import AbstractView from './abstract-view';
+import GameStatView from './game-stat-view';
 import BackButtonView from './back-button-view';
-import {BASE_LEFT_LIVES} from '../data';
 
 export default class HeaderView extends AbstractView {
-  constructor(state) {
+  constructor(state = undefined) {
     super();
-    this.state = state;
+    this.gameStat = new GameStatView(state);
     this.backButton = new BackButtonView();
   }
   get template() {
     return `
       <header class="header">
         ${this.backButton.element.innerHTML}
-        <h1 class="game__timer">${this.state.leftSeconds}</h1>
-        <div class="game__lives">
-          ${new Array(BASE_LEFT_LIVES - this.state.leftLives)
-            .fill(`<img src="img/heart__empty.svg" class="game__heart" alt="Life" width="32" height="32">`)
-            .join(``)}
-          ${new Array(this.state.leftLives)
-              .fill(`<img src="img/heart__full.svg" class="game__heart" alt="Life" width="32" height="32">`)
-              .join(``)}
-        </div>
+        ${this.gameStat.element.innerHTML}
       </header>
     `;
   }
