+++ js/views/rules-view.js
@@ -1,6 +1,6 @@
-import BackButtonView from './back-button-view';
 import AbstractView from './abstract-view';
-import FooterView from './footer';
+import FooterView from './footer-view';
+import HeaderView from "./header-view";
 
 export default class RulesView extends AbstractView {
   constructor(data) {
@@ -9,15 +9,13 @@ export default class RulesView extends AbstractView {
     this.description = data.description;
     this.placeholder = data.placeholder;
     this.button = data.button;
-    this.backButton = new BackButtonView();
+    this.header = new HeaderView();
     this.footer = new FooterView();
   }
 
   get template() {
     return `
-      <header class="header">
-        ${this.backButton.element.innerHTML}
-      </header>
+      ${this.header.element.innerHTML}
       <div class="rules">
         <h1 class="rules__title">${this.title}</h1>
         <p class="rules__description">${this.description}</p>
